Created the taskEncoder task

+++ CanOpener/lib/sensorLib/encoder_counter.h
@@ -42,7 +42,7 @@ public:
     /** @brief   Return the current position count from the timer.
      *  @returns The value in the timer's count register
      */
-    uint16_t getCount (void)
+    uint32_t getCount (void)
     {
         return p_timer->getCount ();
     }
